+++ test/compile/test_var.py
@@ -58,7 +58,7 @@ class TestTypeToStr:
                                "array1d(array2d(0..0, 0..0, slice_3d(arr, [0..0, 0..0, 1..1], 0..0, 0..0, 0..0)))")
                               ])
     def test_flatten(self, v, expected):
-        assert to_str(v, flatten_array=True) == expected
+        assert to_str(v, flatten_arg=True) == expected
 
     @pytest.mark.parametrize("r, expected", [(range(100), "0..99"), (range(1, 10), "1..9"), (range(0, 100), "0..99"),
                                              (range(-10, 10), "-10..9"), (range(-10, -9), "-10..-10")])