+++ changelog.md
@@ -1,9 +1,10 @@
 ### 0.1.2
 #### Fixed
 - some method of Operation and Constraint classes which were
-  accessible by and visible for user are now hidden.
-- error with arithmetic and compare operation are generated now 
-  on python level instead of minizinc level.
+  accessible by and visible for user are now hidden
+#### Added
+- ``increasing`` and ``decreasing`` constraints
+  
   
 ### 0.1.1
 #### Added